Move validation to the root of the server
